# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Ecom API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/user/addtowishlist": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Allows the authenticated user to add a product to their wishlist.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Add to Wishlist",
                "operationId": "AddToWishlist",
                "parameters": [
                    {
                        "description": "Product ID",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.datas"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product successfully added to wishlist.",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "400": {
                        "description": "Invalid input or product error",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
...
        "/user/changepassword": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Change the user's password with a new one.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Change Password",
                "parameters": [
                    {
                        "description": "Old and New Passwords",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.userInputs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/user/listwishlist/{page}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieves the authenticated user's wishlist items.",
                "produces": [
                    "application/json"
                ],
                "summary": "List Wishlist",
                "operationId": "ListWishlist",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Wishlist items",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "400": {
                        "description": "Database error or no products found in wishlist",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
...
        "/user/success": {
            "get": {
                "description": "Displays a success page after payment completion.",
                "produces": [
                    "text/html"
                ],
                "summary": "Payment Success",
                "operationId": "Success",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Payment ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment success page",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "400": {
                        "description": "Error while displaying success page",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
        "/user/updateprofile": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update the user's profile with new information.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Update Profile",
                "parameters": [
                    {
                        "description": "New Profile Information",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.inputDatas"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/userdetail": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get details of the authenticated user including first name, last name, username, email, phone number, and wallet balance.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get User Details",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/user/wishlist/addtocart": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Adds a product from the user's wishlist to the cart.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Add to Cart from Wishlist",
                "operationId": "AddToCartFromWishlist",
                "responses": {
                    "200": {
                        "description": "Products successfully added to the cart",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "400": {
                        "description": "Database or product error",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
        "/user/wishlist/delete/{wishlist_id}": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Removes a product from the user's wishlist.",
                "summary": "Remove from Wishlist",
                "operationId": "RemoveFromWishlist",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Wishlist item ID",
                        "name": "wishlist_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product successfully removed from wishlist.",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "400": {
                        "description": "Integer conversion error or failed to find in wishlist",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "controllers.datas": {
            "type": "object",
            "properties": {
                "productid": {
                    "type": "integer"
                }
            }
        },
        "controllers.inputDatas": {
            "type": "object",
            "properties": {
                "newemail": {
                    "type": "string"
                },
                "newfirstname": {
                    "type": "string"
                },
                "newphone": {
                    "type": "string"
                },
                "newsecondname": {
                    "type": "string"
                },
                "newusername": {
                    "type": "string"
                }
            }
        },
...
        "controllers.userInputs": {
            "type": "object",
            "properties": {
                "newpassword": {
                    "type": "string"
                },
                "oldpassword": {
                    "type": "string"
                }
            }
        },
...
    }
}